third_party/WebKit/LayoutTests/editing/execCommand/remove_format_with_unfocused_selection.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_own_property(window, 'testRunner', 'This test requires testRunner');

  assert_selection(
  [
    '<div contenteditable><b>^text|</b></div>',
    '<a href="http://www.example.com/">link</a>'
  ].join(''),
  selection => {
    const link = selection.document.querySelector('a');
    link.focus();
    testRunner.execCommand('removeFormat');
    assert_equals(selection.document.activeElement, link);
  },
  [
    '<div contenteditable><b>^text|</b></div>',
    '<a href="http://www.example.com/">link</a>'
  ].join(''));
}, 'RemoveFormat with unfocused selection in contenteditable div');
</script>
